feat: extract shared modules — single source of truth for types, config, and constants
Config defaults, trade types (PaperTrade, TradeStatus, TradeMode), auto-trade event
types, status sets (ACTIVE/CLOSED/EXCLUDED), ET date helpers, and currency formatters
were duplicated across auto-trader, frontend, and edge functions causing drift.

Creates shared/ at repo root with pure TS modules imported by all three runtimes.
Also adds dynamic bucket rebalancing (LT allocation → day/swing when suggested finds disabled).
