+++ auto-trader/package.json
@@ -6,7 +6,7 @@
   "type": "module",
   "scripts": {
     "build": "tsc",
-    "start": "node dist/index.js",
+    "start": "node dist/auto-trader/src/index.js",
     "dev": "tsx watch src/index.ts",
     "test": "tsx src/lib/tradePerformanceMetrics.test.ts"
   },
